membenarkan struktur folder dan mengubah beberapa image manjadi svg
