--- src/components/common/ProtectedRoute.jsx
@@ -1,18 +1,0 @@
-import { Navigate } from "react-router-dom";
-import useAuth from "../../lib/hooks/useAuth";
-
-const ProtectedRoute = ({ children }) => {
-  const { isLoggedIn, user, initialized } = useAuth();
-
-  if (!initialized) {
-    return null; 
-  }
-
-  if (!isLoggedIn || !user) {
-    return <Navigate to="/" />;
-  }
-
-  return children;
-};
-
-export default ProtectedRoute;
