+++ src/routes/LobbyUtangga.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Container, Row, Col, Button } from "react-bootstrap";
 import roomUtanggaLogo from "../assets/common/utangga-lobby-logo.png";
 import "../style/routes/LobbyUtangga.css";
-import Header from "../components/common/Header";
+import Header from "../components/Header";
 
 function LobbyUtangga() {
   return (
