+++ src/style/components/Header.css
@@ -1,4 +1,4 @@
-#default-profile-icon {
+/* #default-profile-icon {
    background-color: #fff;
    border-radius: 50%;
    padding: 10px 0px;
@@ -6,5 +6,5 @@
    justify-content: center;
    align-items: center;
    cursor: pointer;
-}
+} */
